refactor : chatroom, notice socket 통합
